Week-9/Day-4/requirements.txt
groq>=0.9.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.26.0
pyautogen>=0.2.35
